[refactor] Moved custom evaluate exceptions to common utils
